backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
Pillow
mss
python-multipart
python-dotenv
pipecat-ai
pipecat-ai[deepgram]
pipecat-ai[openai]
websockets
